+++ src/main/java/chess/model/piece/Bishop.java
@@ -1,12 +1,12 @@
 package chess.model.piece;
 
-import static chess.model.Direction.DOWN_LEFT;
-import static chess.model.Direction.DOWN_RIGHT;
-import static chess.model.Direction.UP_LEFT;
-import static chess.model.Direction.UP_RIGHT;
+import static chess.model.position.Direction.DOWN_LEFT;
+import static chess.model.position.Direction.DOWN_RIGHT;
+import static chess.model.position.Direction.UP_LEFT;
+import static chess.model.position.Direction.UP_RIGHT;
 
-import chess.model.Direction;
-import chess.model.Position;
+import chess.model.position.Direction;
+import chess.model.position.Position;
 
 public class Bishop extends Piece {
