+++ src/main/java/chess/model/piece/Piece.java
@@ -1,6 +1,6 @@
 package chess.model.piece;
 
-import chess.model.Position;
+import chess.model.position.Position;
 
 public abstract class Piece{
