+++ src/main/java/chess/model/piece/Queen.java
@@ -1,16 +1,16 @@
 package chess.model.piece;
 
-import static chess.model.Direction.DOWN;
-import static chess.model.Direction.DOWN_LEFT;
-import static chess.model.Direction.DOWN_RIGHT;
-import static chess.model.Direction.LEFT;
-import static chess.model.Direction.RIGHT;
-import static chess.model.Direction.UP;
-import static chess.model.Direction.UP_LEFT;
-import static chess.model.Direction.UP_RIGHT;
+import static chess.model.position.Direction.DOWN;
+import static chess.model.position.Direction.DOWN_LEFT;
+import static chess.model.position.Direction.DOWN_RIGHT;
+import static chess.model.position.Direction.LEFT;
+import static chess.model.position.Direction.RIGHT;
+import static chess.model.position.Direction.UP;
+import static chess.model.position.Direction.UP_LEFT;
+import static chess.model.position.Direction.UP_RIGHT;
 
-import chess.model.Direction;
-import chess.model.Position;
+import chess.model.position.Direction;
+import chess.model.position.Position;
 
 public class Queen extends Piece {
